chore: refactor ent package structure and update dependencies
- Move ent/driver/multi to ent/multidriver
- Remove entgo.io/ent dependency from root go.mod
- Add ent and multidriver modules with required dependencies
- Update versions.yaml to include new ent-related packages
