+++ ent/go.mod
@@ -1,0 +1,15 @@
+module github.com/go-fries/fries/ent/v3
+
+go 1.22.10
+
+require (
+	github.com/go-kratos/kratos/v2 v2.8.3
+	github.com/stretchr/testify v1.10.0
+)
+
+require (
+	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/kr/text v0.2.0 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
+)
