+++ ent/multidriver/go.mod
@@ -1,0 +1,15 @@
+module github.com/go-fries/fries/multidriver/v3
+
+go 1.22.10
+
+require (
+	entgo.io/ent v0.14.0
+	github.com/stretchr/testify v1.10.0
+)
+
+require (
+	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/google/uuid v1.3.0 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
+)
